+++ src/app/(whitebg)/users/seeds/_components/(tabs)/TabView/index.tsx
@@ -6,13 +6,14 @@ import HistoryTab from '../HistoryTab';
 import TabButtonWrapper from '@/components/TabButton/TabButtonWrapper';
 import TabButton from '@/components/TabButton';
 import BuyTab from '../BuyTab';
+import styles from './TabView.module.css';
 
 const TabView = () => {
   const router = useRouter();
   const tab = useCurrentParams('tabs', 'buy');
 
   return (
-    <>
+    <div className={styles.wrapper}>
       <TabButtonWrapper>
         {['buy', 'history'].map((t) => (
           <TabButton key={t} onClick={() => router.push(`?tabs=${t}`)} active={tab === t}>
@@ -21,11 +22,9 @@ const TabView = () => {
         ))}
       </TabButtonWrapper>
 
-      <div>
-        {tab === 'buy' && <BuyTab />}
-        {tab === 'history' && <HistoryTab />}
-      </div>
-    </>
+      {tab === 'buy' && <BuyTab />}
+      {tab === 'history' && <HistoryTab />}
+    </div>
   );
 };
